requirements.txt
numpy==1.26.4
opencv-python==4.10.0.82
cryptography==43.0.0
Pillow==10.2.0
